requirements.txt
mysql.connector
cryptography.fernet
